requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20210324.2=h9c3ff4c_0
absl-py=0.15.0=pyhd8ed1ab_0
aiohttp=3.8.3=py38h0a891b7_0
aiosignal=1.2.0=pyhd8ed1ab_0
alsa-lib=1.2.3.2=h166bdaf_0
argcomplete=2.0.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.5=pyhd8ed1ab_0
boto=2.49.0=pypi_0
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.10.11=h06a4308_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2022.9.24=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py38h578d9bd_0
cloudpickle=2.2.0=pypi_0
crcmod=1.7=pypi_0
cryptography=37.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cudnn=8.2.1.32=h86fa8c9_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
dm-tree=0.1.7=pypi_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
fasteners=0.18=pypi_0
ffmpeg=4.3.2=h37c90e5_3
fftw=3.3.10=nompi_h77c792f_102
fontconfig=2.14.0=h8e229c2_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.1=py38h0a891b7_0
gast=0.4.0=pyh9f0ad1d_0
gcs-oauth2-boto-plugin=3.0=pypi_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
glib=2.68.4=h9c3ff4c_0
glib-tools=2.68.4=h9c3ff4c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-apitools=0.5.32=pypi_0
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
google-reauth=0.1.1=pypi_0
graphite2=1.3.14=h295c915_1
grpc-cpp=1.39.1=h850795e_1
grpcio=1.39.0=py38hdd6454d_0
gst-plugins-base=1.18.5=hf529b03_0
gstreamer=1.18.5=h76c114f_0
gsutil=5.14=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h5py=3.1.0=nompi_py38hafa665b_100
harfbuzz=2.9.1=h83ec7ef_1
hdf5=1.10.6=nompi_h6a2412b_1114
httplib2=0.20.4=pypi_0
icu=68.2=h9c3ff4c_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=5.0.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py38h06a4308_0
ipython=8.4.0=py38h06a4308_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.18.1=py38h06a4308_1
jpeg=9e=h7f8727e_0
jsonschema=4.16.0=py38h06a4308_0
jupyter_client=7.3.5=py38h06a4308_0
jupyter_core=4.11.1=py38h06a4308_0
keras=2.6.0=pyhd8ed1ab_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
learn2learn=0.1.7=pypi_0
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libcurl=7.85.0=h7bff187_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.3=he6710b0_2
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libglib=2.68.4=h3e27bee_0
libgomp=12.1.0=h8d9b700_16
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm11=11.1.0=hf817b99_3
libnghttp2=1.47.0=h727a467_0
libogg=1.3.5=h27cfd23_1
libopencv=4.5.3=py38h5627943_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libpq=13.5=hd57d9b9_1
libprotobuf=3.16.0=h780b84a_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.12=h166bdaf_3
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py38h0a891b7_1
matplotlib-inline=0.1.6=py38h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
monotonic=1.6=pypi_0
multidict=6.0.2=py38h0a891b7_1
mysql-common=8.0.29=haf5c9bc_1
mysql-libs=8.0.29=h28c427c_1
nbformat=5.5.0=py38h06a4308_0
nccl=2.12.12.1=h0800d71_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py38h06a4308_0
nettle=3.7.3=hbbd107a_1
nspr=4.33=h295c915_0
nss=3.77=h2350873_0
numexpr=2.8.3=py38h807cd23_0
numpy=1.19.5=py38h9894fe3_2
numpy-base=1.23.1=py38ha15fc14_0
oauth2client=4.1.3=pypi_0
oauthlib=3.2.1=pyhd8ed1ab_0
opencv-python-headless=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py38h6a678d5_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=9.3.0=pypi_0
pip=22.2.2=py38h06a4308_0
pixman=0.40.0=h36c2ea0_0
pkgutil-resolve-name=1.3.10=py38h06a4308_0
plotly=5.9.0=py38h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.16.0=py38h709712a_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-opencv=4.5.3=py38he5a9106_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.5.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38h06a4308_0
python-flatbuffers=1.12=pyhd8ed1ab_1
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py38h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py38h6a678d5_0
qpth=0.0.15=pypi_0
qt=5.12.9=hda022c4_4
re2=2021.09.01=h9c3ff4c_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
retry-decorator=1.1.1=pypi_0
rsa=4.7.2=pypi_0
scipy=1.9.1=py38h14f4228_0
setuptools=63.4.1=py38h06a4308_0
six=1.15.0=pyh9f0ad1d_0
snappy=1.1.9=hbd366e4_1
sqlite=3.39.3=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
tenacity=8.0.1=py38h06a4308_1
tensorboard=2.6.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py38h2b5fc30_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.6.0=cuda112py38hbe5352d_2
tensorflow-base=2.6.0=cuda112py38heae9c4c_2
tensorflow-estimator=2.6.0=cuda112py38hb2194ef_2
tensorflow-gpu=2.6.0=cuda112py38h0bbbad9_2
tensorflow-model-optimization=0.7.3=pypi_0
termcolor=1.1.0=pyhd8ed1ab_3
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py38_cu113
torchsummary=1.5.1=pypi_0
torchvision=0.13.1=py38_cu113
tornado=6.2=py38h5eee18b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.11=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.12.1=py38h497a2fe_3
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h5eee18b_0
yarl=1.8.1=py38h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.9.0=pyhd8ed1ab_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
